♻️ refactor(login): move login page into feature module
Move the login page implementation, styles, login mode resolver, and tests under src/features/login.

Keep src/pages/LoginPage.tsx as a route wrapper so login behavior now lives in the feature layer.
