+++ src/features/login/LoginPage.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/variables.scss' as *;
+@use '../../styles/variables.scss' as *;
 
 // 主容器 - 左右分栏布局
 .container {
